Fix bug where calling setImmortal on a raw sprite could cause corruption
Corruption would only occur if we attempted to setImmortal() a raw sprite
before we had ever rendered it.
get() in SpriteManager handles raw sprites with directInsert, but calling
setImmortal would skip straight to SpriteCache::setImmortal, which ensures
the sprite exists by calling get() and then setting the immortal flag.
However, since this was implemented inside SpriteCache, it wasn't handling
raw sprites, so it woul give an invalid sprite cache lookup error, and
corrupt the mUsedList member. The solution is to call get() in
SpriteManager::setImmortal before calling SpriteCache::SetImmortal, to
ensure that the raw sprite is created before we go into SpriteCache.