fix bug with hit animation
an attack is considered "over" when the attack animation is complete.
This was being checked by comparing the actor's current animation
to the attack animation. However, if the actor is hit during the attack
animation, they play the hit animation, which marks the attack as done.
This lets them attack again immediately. Here, I change d this behaviour,
so the attack will resume after the hit anim, and you can't attack again
until it's really finished.