server waits for clients + verification mode
Server can turn on a "full verify" flag, which will do a full map save
on every tick and send it to the client along with the normal update.
The client will then save after every tick and verify that the saved data
is exactly the same as the save received from the server.