Optimised FixedPoint
Replaced the muldiv implementation. It was using a library to fake full
128-bit operations, but it turns out the AMD64 instruction set can do the
subset of operations that we need just using its normal integer multiply
and divide instructions.

Tested the performance improvement in a pretty crude fashion, but it shows
there are significant gains. I just ran the unit tests on Linux and
timed the execution.

Release: 35% drop (0.166s -> 0.107s)
Debug: 69% drop (0.676s -> 0.207s)