Fix animation stuttering when walking
The problem was that sometimes in a single tick we would overshoot the
next tile. This would cause us to "arrive" and stop, until in the next
tick the movement handler would tell us to continue to the next tile in
our path.
The solution is to return the amount of movement we have left, and keep
looping the update function until we've used it all up.