Fix alignment of uniform buffers
We are packing separate bindings of uniform blocks (Vertex and Fragment)
into one buffer, and before we were just using a struct containing both
as members, then dumping that into the GPU buffer. We could then use
offsetof() to get the offsets for binding.
The problem, is that in opengl (and probably other apis) a uniform block
binding cannot just be at any old offset inside a buffer. We need to
query at runtime for GL_UNIFORM_BUFFER_OFFSET_ALIGNMENT.

So, this commit adds a wrapper around a vector of bytes, which allows us
to create a block of cpu memory with the right size and alignment.