Changes to positioning to fix missiles
Entities logical position is now centred on the tile they occupy.
Effectively an entity at tile x,y now has fractional position x+0.5,y+0.5.
This makes missiles work much better, as before you would spawn a
missile at the shooters position, and it would start at the top of the tile
the shooter was occupying, and move towards the top of the tile the target
was occupying.
Now it will start at the centre of the shooter's tile and move towards the
centre of the target's tile, so it is guaranteed to hit. A side bonus is
that the logic is easier to understand too, imo.